Simplificação do código do back.
