+++ front/.env
@@ -1,2 +1,4 @@
-VITE_URL_HOST="localhost" #Importante trocar para o IP da máquina que está rodando o back para acessar de outros dispositivos
+#Importante trocar para o IP da máquina que está hospedando para acessar
+#de outros dispositivos ou o localhost para acessar localmente
+VITE_URL_HOST="192.168.2.169" 
 VITE_PORT=3000
